删去了父类User
