+++ xyw/src/rj7/bean/Admin.java
@@ -1,11 +1,32 @@
 package rj7.bean;
 
-public class Admin extends User {
+public class Admin {
 
+	private String id;
+	private String username;
+	private String pswd;
 	private String phone;
 	private String email;
-	private String level; // 管理员级别
+	private String level; 
 	
+	public String getId() {
+		return id;
+	}
+	public void setId(String id) {
+		this.id = id;
+	}
+	public String getUsername() {
+		return username;
+	}
+	public void setUsername(String username) {
+		this.username = username;
+	}
+	public String getPswd() {
+		return pswd;
+	}
+	public void setPswd(String pswd) {
+		this.pswd = pswd;
+	}
 	public String getPhone() {
 		return phone;
 	}
