--- xyw/src/rj7/bean/User.java
@@ -1,28 +1,0 @@
-package rj7.bean;
-
-public class User {
-
-	private String id;
-	private String username;
-	private String pswd;
-	
-	public String getId() {
-		return id;
-	}
-	public void setId(String id) {
-		this.id = id;
-	}
-	public String getUsername() {
-		return username;
-	}
-	public void setUsername(String username) {
-		this.username = username;
-	}
-	public String getPswd() {
-		return pswd;
-	}
-	public void setPswd(String pswd) {
-		this.pswd = pswd;
-	}
-	
-}
